# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Unified Operations Platform API
  description: |
    Comprehensive API for the Unified Operations Platform supporting event management,
    resource planning, financial operations, and multi-tenant workflows.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: Proprietary
    url: https://example.com/license
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /users/{userId}:
    get:
      summary: Get user
      description: Retrieve a specific user by ID
      operationId: getUser
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      summary: Update user
      description: Update an existing user
      operationId: updateUser
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete user
      description: Delete a user (soft delete)
      operationId: deleteUser
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "204":
          description: User deleted successfully
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /events:
    get:
      summary: List events
      description: Retrieve a paginated list of events
      operationId: listEvents
      tags:
        - Events
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, confirmed, in_progress, completed, cancelled]
          description: Filter by event status
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: Filter events starting after this date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: Filter events ending before this date
        - name: venueId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by venue
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"

    post:
      summary: Create event
      description: Create a new event
      operationId: createEvent
      tags:
        - Events
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{eventId}:
    get:
      summary: Get event
      description: Retrieve a specific event by ID
      operationId: getEvent
      tags:
        - Events
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/EventIdParam"
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      summary: Update event
      description: Update an existing event
      operationId: updateEvent
      tags:
        - Events
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/EventIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventRequest"
      responses:
        "200":
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

    delete:
      summary: Delete event
      description: Delete an event (soft delete)
      operationId: deleteEvent
      tags:
        - Events
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/EventIdParam"
      responses:
        "204":
          description: Event deleted successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    SearchParam:
      name: search
      in: query
      schema:
        type: string
      description: Search query string
    UserIdParam:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User ID
...
    EventIdParam:
      name: eventId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Event ID
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        avatar:
          type: string
          format: uri
        role:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
        organizationId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [draft, confirmed, in_progress, completed, cancelled]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        venueId:
          type: string
          format: uuid
        venue:
          $ref: "#/components/schemas/Venue"
        capacity:
          type: integer
        budget:
          type: number
          format: decimal
        organizationId:
          type: string
          format: uuid
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateEventRequest:
      type: object
      required:
        - name
        - startDate
        - endDate
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        venueId:
          type: string
          format: uuid
        capacity:
          type: integer
        budget:
          type: number
          format: decimal
    UpdateEventRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [draft, confirmed, in_progress, completed, cancelled]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        venueId:
          type: string
          format: uuid
        capacity:
          type: integer
        budget:
          type: number
          format: decimal
    EventListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    Venue:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postalCode:
          type: string
        capacity:
          type: integer
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
            format: uri
        status:
          type: string
          enum: [active, inactive, maintenance]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []